requirements.txt
Django==2.2.4
djangorestframework==3.10.2
psycopg2-binary==2.8.3
shapely==1.6.4